Move encoding/decoding errors to internal package
